+++ src/libraries/Common/src/Interop/Unix/System.Security.Cryptography.Native/Interop.EvpPkey.cs
@@ -321,214 +321,6 @@ internal static partial class Interop
             }
         }
 
-        /// <summary>
-        /// Returns the OID as ASN1_OBJECT pointer.
-        /// </summary>
-        [LibraryImport(Libraries.CryptoNative, StringMarshalling = StringMarshalling.Utf8)]
-        private static partial IntPtr CryptoNative_EvpPKeyGetEcGroupOid(SafeEvpPKeyHandle pkey);
-
-        /// <summary>
-        /// Returns the OID as a string or null if curve is not named.
-        /// </summary>
-        internal static string? EvpPKeyGetEcGroupName(SafeEvpPKeyHandle pkey)
-        {
-            IntPtr oid = CryptoNative_EvpPKeyGetEcGroupOid(pkey);
-
-            if (oid == IntPtr.Zero)
-            {
-                return null;
-            }
-
-            return GetOidValue(oid);
-        }
-
-        [LibraryImport(Libraries.CryptoNative)]
-        private static partial int CryptoNative_EvpPKeyGetEcKeyParameters(
-            SafeEvpPKeyHandle key,
-            [MarshalAs(UnmanagedType.Bool)] bool includePrivate,
-            out SafeBignumHandle qx_bn, out int x_cb,
-            out SafeBignumHandle qy_bn, out int y_cb,
-            out SafeBignumHandle d_bn, out int d_cb);
-
-        internal static ECParameters EvpPKeyGetEcKeyParameters(
-            SafeEvpPKeyHandle key,
-            bool includePrivate)
-        {
-            SafeBignumHandle qx_bn, qy_bn, d_bn;
-            int qx_cb, qy_cb, d_cb;
-            ECParameters parameters = default;
-
-            int rc = CryptoNative_EvpPKeyGetEcKeyParameters(
-                key,
-                includePrivate,
-                out qx_bn, out qx_cb,
-                out qy_bn, out qy_cb,
-                out d_bn, out d_cb);
-
-            using (qx_bn)
-            using (qy_bn)
-            using (d_bn)
-            {
-                if (rc == -1)
-                {
-                    throw new CryptographicException(SR.Cryptography_CSP_NoPrivateKey);
-                }
-                else if (rc != 1)
-                {
-                    throw Interop.Crypto.CreateOpenSslCryptographicException();
-                }
-
-                // Match Windows semantics where qx, qy, and d have same length
-                int expectedSize = GetEvpPKeySizeBytes(key);
-                int cbKey = GetMax(qx_cb, qy_cb, d_cb);
-
-                Debug.Assert(
-                    cbKey <= expectedSize,
-                    $"Expected output size was {expectedSize}, which a parameter exceeded. qx={qx_cb}, qy={qy_cb}, d={d_cb}");
-
-                cbKey = GetMax(cbKey, expectedSize);
-
-                parameters.Q = new ECPoint
-                {
-                    X = Crypto.ExtractBignum(qx_bn, cbKey),
-                    Y = Crypto.ExtractBignum(qy_bn, cbKey)
-                };
-                parameters.D = d_cb == 0 ? null : Crypto.ExtractBignum(d_bn, cbKey);
-            }
-
-            return parameters;
-        }
-
-        [LibraryImport(Libraries.CryptoNative)]
-        private static unsafe partial void CryptoNative_BufferFree(byte* ptr);
-
-        [LibraryImport(Libraries.CryptoNative)]
-        private static unsafe partial int CryptoNative_EvpPKeyGetEcCurveParameters(
-            SafeEvpPKeyHandle key,
-            [MarshalAs(UnmanagedType.Bool)] bool includePrivate,
-            out ECCurve.ECCurveType curveType,
-            out SafeBignumHandle qx, out int x_cb,
-            out SafeBignumHandle qy, out int y_cb,
-            out SafeBignumHandle d, out int d_cb,
-            out SafeBignumHandle p, out int P_cb,
-            out SafeBignumHandle a, out int A_cb,
-            out SafeBignumHandle b, out int B_cb,
-            out SafeBignumHandle gx, out int Gx_cb,
-            out SafeBignumHandle gy, out int Gy_cb,
-            out SafeBignumHandle order, out int order_cb,
-            out SafeBignumHandle cofactor, out int cofactor_cb,
-            out byte* seed, out int seed_cb);
-
-        internal static unsafe ECParameters EvpPKeyGetEcCurveParameters(
-            SafeEvpPKeyHandle key,
-            bool includePrivate)
-        {
-            ECCurve.ECCurveType curveType;
-            SafeBignumHandle qx_bn, qy_bn, d_bn, p_bn, a_bn, b_bn, gx_bn, gy_bn, order_bn, cofactor_bn;
-            byte* seed = null;
-            int qx_cb, qy_cb, p_cb, a_cb, b_cb, gx_cb, gy_cb, order_cb, cofactor_cb, seed_cb, d_cb;
-            try
-            {
-                int rc = CryptoNative_EvpPKeyGetEcCurveParameters(
-                    key,
-                    includePrivate,
-                    out curveType,
-                    out qx_bn, out qx_cb,
-                    out qy_bn, out qy_cb,
-                    out d_bn, out d_cb,
-                    out p_bn, out p_cb,
-                    out a_bn, out a_cb,
-                    out b_bn, out b_cb,
-                    out gx_bn, out gx_cb,
-                    out gy_bn, out gy_cb,
-                    out order_bn, out order_cb,
-                    out cofactor_bn, out cofactor_cb,
-                    out seed, out seed_cb);
-
-                using (qx_bn)
-                using (qy_bn)
-                using (d_bn)
-                using (p_bn)
-                using (a_bn)
-                using (b_bn)
-                using (gx_bn)
-                using (gy_bn)
-                using (order_bn)
-                using (cofactor_bn)
-                {
-                    if (rc == -1)
-                    {
-                        throw new CryptographicException(SR.Cryptography_CSP_NoPrivateKey);
-                    }
-                    else if (rc != 1)
-                    {
-                        throw Interop.Crypto.CreateOpenSslCryptographicException();
-                    }
-
-                    int cbFieldLength;
-                    int pFieldLength;
-                    if (curveType == ECCurve.ECCurveType.Characteristic2)
-                    {
-                        // Match Windows semantics where a,b,gx,gy,qx,qy have same length
-                        // Treat length of m separately as it is not tied to other fields for Char2 (Char2 not supported by Windows)
-                        cbFieldLength = GetMax(new[] { a_cb, b_cb, gx_cb, gy_cb, qx_cb, qy_cb });
-                        pFieldLength = p_cb;
-                    }
-                    else
-                    {
-                        // Match Windows semantics where p,a,b,gx,gy,qx,qy have same length
-                        cbFieldLength = GetMax(new[] { p_cb, a_cb, b_cb, gx_cb, gy_cb, qx_cb, qy_cb });
-                        pFieldLength = cbFieldLength;
-                    }
-
-                    // Match Windows semantics where order and d have same length
-                    int cbSubgroupOrder = GetMax(order_cb, d_cb);
-
-                    // Copy values to ECParameters
-                    ECParameters parameters = default;
-                    parameters.Q = new ECPoint
-                    {
-                        X = Crypto.ExtractBignum(qx_bn, cbFieldLength),
-                        Y = Crypto.ExtractBignum(qy_bn, cbFieldLength)
-                    };
-                    parameters.D = d_cb == 0 ? null : Crypto.ExtractBignum(d_bn, cbSubgroupOrder);
-
-                    var curve = parameters.Curve;
-                    curve.CurveType = curveType;
-                    curve.A = Crypto.ExtractBignum(a_bn, cbFieldLength)!;
-                    curve.B = Crypto.ExtractBignum(b_bn, cbFieldLength)!;
-                    curve.G = new ECPoint
-                    {
-                        X = Crypto.ExtractBignum(gx_bn, cbFieldLength),
-                        Y = Crypto.ExtractBignum(gy_bn, cbFieldLength)
-                    };
-                    curve.Order = Crypto.ExtractBignum(order_bn, cbSubgroupOrder)!;
-
-                    if (curveType == ECCurve.ECCurveType.Characteristic2)
-                    {
-                        curve.Polynomial = Crypto.ExtractBignum(p_bn, pFieldLength)!;
-                    }
-                    else
-                    {
-                        curve.Prime = Crypto.ExtractBignum(p_bn, pFieldLength)!;
-                    }
-
-                    // Optional parameters
-                    curve.Cofactor = cofactor_cb == 0 ? null : Crypto.ExtractBignum(cofactor_bn, cofactor_cb);
-                    Span<byte> seedSpan = new Span<byte>(seed, seed_cb);
-                    curve.Seed = seedSpan.ToArray();
-
-                    parameters.Curve = curve;
-                    return parameters;
-                }
-            }
-            finally
-            {
-                if (seed != null)
-                    CryptoNative_BufferFree(seed);
-            }
-        }
-
         internal enum EvpAlgorithmId
         {
             Unknown = 0,
