+++ src/libraries/System.Net.Security/tests/FunctionalTests/TestConfiguration.cs
@@ -10,8 +10,11 @@ using Xunit;
 
 namespace System.Net.Security.Tests
 {
-    internal static partial class TestConfiguration
+    internal static class TestConfiguration
     {
+        public const int PassingTestTimeoutMilliseconds = 1 * 60 * 1000;
+        public static TimeSpan PassingTestTimeout => TimeSpan.FromMilliseconds(PassingTestTimeoutMilliseconds);
+
         public const string Realm = "TEST.COREFX.NET";
         public const string KerberosUser = "krb_user";
         public const string DefaultPassword = "password";
