+++ src/native/libs/System.Security.Cryptography.Native/entrypoints.c
@@ -171,9 +171,8 @@ static const Entry s_cryptoNative[] =
     DllImportEntry(CryptoNative_EvpPkeySetEcKey)
     DllImportEntry(CryptoNative_EvpPKeyBits)
     DllImportEntry(CryptoNative_EvpPKeyGetEcKeyParameters)
-    DllImportEntry(CryptoNative_EvpPKeyGetEcGroupOid)
+    DllImportEntry(CryptoNative_EvpPKeyGetEcGroupNid)
     DllImportEntry(CryptoNative_EvpPKeyGetEcCurveParameters)
-    DllImportEntry(CryptoNative_BufferFree)
     DllImportEntry(CryptoNative_EvpRC2Cbc)
     DllImportEntry(CryptoNative_EvpRC2Ecb)
     DllImportEntry(CryptoNative_EvpSha1)
